interface/isl_test_cpp-noexceptions.cc: extract out generic tests
In particular, extract out those tests that do not depend
on the absence of exceptions.
These will be reused to test the C++ bindings with exceptions,
which will be added next.
