+++ Makefile.am
@@ -391,6 +391,7 @@ EXTRA_DIST = \
 	interface/isl-noexceptions.h.pre \
 	interface/isl-noexceptions.h.bot \
 	interface/isl.py.top \
+	interface/isl_test_cpp-generic.cc \
 	interface/isl_test_python.py \
 	test_inputs
